Add proper support for fields whose type is an Enum
This commit adds support for message fields whose type is an Enum.
Among other things, this commit deals with:

* Enum fields not being nullable (strangely, `nullable` `false` is not an
option even if it is the only option available.
* Enum fields needing a casttype option.
* Enum fields needing extra gogo options.

Also, this commit update the example to make it aware of enums and refactors it
so it can be run in the CI server.