Improve IsNullable and gogoproto.nullable support
This commit tests IsNullable for:

* protobuf.Named
* protobuf.Alias
* protobuf.Basic
* protobuf.Map

And fixes the implementation for Map. A map is not nullable if the type
of their values is not nullable.

This also generates correctly the nullable option for maps and fixes
alias resolution: When an alias was found, the underlying was not
resolve which could cause some messages not to be generated.